Rework the way unique id hashing works.
We need unique identifiers for each font stored in the database so that
we don't have more than one instance of the same font. I was doing this
before by making hashes of the file contents and names. That's fragile
(a single by difference will result in a different hash) and it only
works for OTFs. Instead of doing that, I now create a hash of a sorted
string containing every coordinate used in the font's outline and the
number of times that a point is placed in that spot.